+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.0.0+3
+* Add `size_t`, `ptrdiff_t`.
+* Add `invoke` to `Expression`.
+
 # 1.0.0+2
 * Allow the creation of `Enum`s.
 * Add `extern`, `inline`, and `enum$` prefixes to `CType`.